[vpdq] Fix macOS build
Clang doesn't like the forward declaration of the templates.
